requirements.txt
mypy>=0.76
flake8>=3.7.9
pylint>=2.4.4
pytest>=5.3.5
requests>=2.21
pillow>=7.0
jsonschema>=3.0.1
beautifulsoup4>=4.8.2
tinydb>=3.15.2
Pillow >= 8.0.0
matplotlib >= 3.3.2
opencv-python >= 4.4.0
tensorflow >= 2.3.1
keras >= 2.4.3
numpy == 1.18.5
h5py == 2.10.0
scikit-learn == 0.24.2